+++ front/locale.php
@@ -55,7 +55,7 @@ global $CFG_GLPI, $TRANSLATE;
 
 // Get messages from translator component
 $messages = $TRANSLATE->getAllMessages($_GET['domain']);
-if (!($messages instanceof \Zend\I18n\Translator\TextDomain)) {
+if (!($messages instanceof \Laminas\I18n\Translator\TextDomain)) {
    Toolbox::logError(sprintf('Unable to get messages from domain "%s".', $_GET['domain']));
    exit('[]');
 }